epub-maker: split epub builder into serializer, docs, and zip modules
